backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
python-multipart==0.0.6
# Machine Learning
tensorflow>=2.13.0
Pillow>=10.0.0
numpy>=1.24.0